Hi there
Because of the great work of our new developer GateKeeper I am happy to
make this commit. This will be the first and the last major WOL commit I
will probably ever make as GateKeeper will take care of this fro now on,
hopefully for a looong time :)

This commit adds WOL support as from GateKeeper's last patch version.
Changes relative to GK's patch were:
- renamed wol_log parameter to wol_long (longitude, not logitude :))
- removed some unused vars from handle_irc.c
- removed the "wolalphaname" bnetd.conf option and use servername
instead
- fixed conf_setdef_wol_addrs() to use WOL_ADDRS defined in
setup_before.h and as such fix a bug where pvpgn did not started if
wol_addrs are commented in bnetd.conf (the default now)
- added EXPERIMENTAL notices to bnetd.conf WOL section
- removed duplicated bnetd.conf.win32 WOL section at the end of the file

Until BerliOS adds "diffstat" command on their CVS servers (so I make
the diff mails use it) we should also include a diffstat output of the
"cvs diff" command of large commits (like this one). Diffstat output for
this commit follows:

 CREDITS                   |    4
 NEWS                      |    6
 UPDATE                    |    5
 conf/bnetd.conf.in        |   21 +
 conf/bnetd.conf.win32     |   21 +
 conf/channel.conf         |    5
 src/bnetd/account_wrap.c  |   15
 src/bnetd/account_wrap.h  |    7
 src/bnetd/channel.c       |  143 ++++++
 src/bnetd/channel.h       |   29 +
 src/bnetd/connection.c    |  192 +++++++++
 src/bnetd/connection.h    |   35 +
 src/bnetd/handle_irc.c    |  958 ++++++++++++++++++++++++++++++++++++----------
 src/bnetd/irc.c           |  163 +++++--
 src/bnetd/message.c       |    5
 src/bnetd/message.h       |   10
 src/bnetd/prefs.c         |  113 +++++
 src/bnetd/prefs.h         |    8
 src/bnetd/server.c        |   16
 src/bnetd/server.h        |    1
 src/common/irc_protocol.h |   81 ++-
 src/common/setup_before.h |    2
 src/common/tag.c          |   10
 src/common/tag.h          |    5
 version-history.txt       |    4
 25 files changed, 1595 insertions(+), 264 deletions(-)